+++ src/helpers.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+use Zarbin\Seo\ZarbinSeo;
+
+if (! function_exists('seo')) {
+    function seo(): ZarbinSeo
+    {
+        return app('zarbin-seo');
+    }
+}